Fuji: - Added MFCopyMemory, MFZeroMemory, MFSetMemory and MFMemCompare. - Replaced CRT functions with the fuji versions.
Caelestis:
- Added Switch action.
- Added SelectionDialog boxes.
- Added pointer finger icon.
- Fixed some bugs.

git-svn-id: https://fuji.googlecode.com/svn/trunk@383 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Include/MFHeap.h
@@ -95,12 +95,11 @@ void operator delete(void *pMemory);
 void operator delete[](void *pMemory);
 #endif
 
-/*
-void* operator new(unsigned int size, void *pMem);
-void* operator new[](unsigned int size, void *pMem);
+// placement new operators
+void* operator new(size_t size, void *pMem);
+void* operator new[](size_t size, void *pMem);
 void operator delete(void *pMemory, void *pMem);
 void operator delete[](void *pMemory, void *pMem);
-*/
 #endif
 
 /**
